refactor: modularize PlaceholderEditor
