+++ src/components/dialogs/index.tsx
@@ -3,7 +3,7 @@ import React from 'react';
 import { DialogProvider as DialogContextProvider } from './DialogContext';
 import { TemplateDialog } from './templates/TemplateDialog';
 import { FolderDialog } from './templates/FolderDialog';
-import { PlaceholderEditor } from './templates/PlaceHolderEditor';
+import { PlaceholderEditor } from './templates/PlaceholderEditor';
 import { AuthDialog } from './auth/AuthDialog';
 import { SettingsDialog } from './settings/SettingsDialog';
 import { ConfirmationDialog } from './common/ConfirmationDialog';
@@ -53,7 +53,7 @@ export { DialogContextProvider } from './DialogContext';
 export { useDialog, useDialogManager } from './DialogContext';
 export { TemplateDialog } from './templates/TemplateDialog';
 export { FolderDialog } from './templates/FolderDialog';
-export { PlaceholderEditor } from './templates/PlaceHolderEditor';
+export { PlaceholderEditor } from './templates/PlaceholderEditor';
 export { AuthDialog } from './auth/AuthDialog';
 export { SettingsDialog } from './settings/SettingsDialog';
 export { ConfirmationDialog } from './common/ConfirmationDialog';
